requirement.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
kyber-py==0.2.2
MarkupSafe==2.1.5
Naked==0.1.32
numpy==2.1.1
packaging==24.1
pluggy==1.5.0
pycryptodome==3.14.1
pytest==8.3.3
pytest-html==4.1.1
pytest-metadata==3.1.1
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
urllib3==2.2.3
